+++ simple-posts-ticker.php
@@ -3,7 +3,7 @@
  * Plugin Name: Simple Posts Ticker
  * Plugin URI: https://wordpress.org/plugins/simple-posts-ticker/
  * Description: The Simple Posts Ticker plugin is a small tool that shows your most recent posts in a marquee style.
- * Version: 1.1.1
+ * Version: 1.1.2
  * Author: Sayan Datta
  * Author URI: https://about.me/iamsayan
  * License: GPLv3
@@ -35,7 +35,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'SPT_PLUGIN_VERSION', '1.1.1' );
+define( 'SPT_PLUGIN_VERSION', '1.1.2' );
 
 // debug scripts
 //define( 'SPT_PLUGIN_ENABLE_DEBUG', 'true' );